% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/patches.R
\name{alpha_var}
\alias{alpha_var}
\title{Calculate alpha variability from community matrix}
\usage{
alpha_var(D, id_var, measurement_var, time_var)
}
\arguments{
\item{D}{Dataframe with time series of species
level measurements of abundance across multiple patches.}

\item{id_var}{Column name within dataframe that identifies unique patches.}

\item{measurement_var}{Column name within dataframe that identifies the
measurement of abundance used.}

\item{time_var}{Column name within dataframe for the temporal identifier.}
}
\description{
Calculate alpha variability from community matrix
}
\author{
Andrew Tredennick
}

