\name{patches-package}
\alias{patches-package}
\alias{patches}
\docType{package}
\title{
Collection of hot pacthes for R
}
\description{
Collection of hot pacthes for R
}
\details{
\tabular{ll}{
Package: \tab patches\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2010-11-29\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

A faster version of \link[base]{sequence} is installed
in the base namespace when patches loads. 

}
\author{
Romain Francois

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\keyword{ package }
