\name{patches-package}
\alias{patches-package}
\alias{patches}
\docType{package}
\title{
Collection of hot pacthes for R
}
\description{
Collection of hot pacthes for R
}
\details{
\tabular{ll}{
Package: \tab patches\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 2010-11-29\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

A faster version of \link[base]{sequence} is installed
in the base namespace when patches loads. The patch is based on a variation 
of the code submitted to R-devel in this thread 
\url{http://article.gmane.org/gmane.comp.lang.r.devel/26157}

More patches will be added

}
\author{
Romain Francois

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\keyword{ package }
